Add search function
Kinda working. Has some bugs ":D
I plan to fix it later, but rn I have other stuff to do.

+++ Models/WordPosition.cs
@@ -2,6 +2,6 @@
 
 public struct WordPosition
 {
-    public long lineIndex;
+    public int lineIndex;
     public long lineOffset;
 }
